flask_auth_service/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
PyJWT==2.8.0
gunicorn==21.2.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
cryptography==42.0.5
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
# Add other dependencies as needed, e.g., for database interaction if storing users here
# cryptography>=3.4.0 # If advanced crypto is needed beyond PyJWT 